Move account work inside lib
