+++ client/src/lib/account/CurrentUserContext.js
