+++ client/src/lib/account/SignInModal.js
@@ -3,7 +3,7 @@ import React, { useRef, useState } from 'react'
 import StandardModal from './StandardModal';
 import StatefulButton from './StatefulButton';
 import SignInForm from './SignInForm';
-import { useEnvironment } from '../lib/BrandingContext';
+import { useEnvironment } from '../BrandingContext';
 
 function SignInModal({
   toggle,
