+++ client/src/lib/index.js
@@ -15,3 +15,9 @@ export { default as FullscreenButton } from './FullscreenButton';
 export { default as Spinner } from './Spinner';
 export { default as UnauthorizedError } from './UnauthorizedError';
 export { default as useEventListener } from './useEventListener';
+
+// Export a subset of the account modules.
+export {
+  Context as CurrentUserContext,
+  Provider as CurrentUserProvider,
+} from './account/CurrentUserContext';
